pyproject.toml
[project]
name = "azurevmmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.21.0",
    "azure-mgmt-compute>=34.1.0",
    "azure-mgmt-network>=28.1.0",
    "azure-mgmt-resource>=23.3.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
]
